requirements.txt
attrs==23.1.0
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
Flask==3.0.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
outcome==1.3.0.post0
paho-mqtt==1.6.1
PySocks==1.7.1
PyYAML==6.0.1
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
